+++ changes.md
@@ -1,5 +1,6 @@
 ##### unreleased
 
+-   Added driver admin. [#6]
 -   Added nav. [#5]
 -   Added team admin. [#4]
 -   Added entities. [#3]